+++ src/DetailsView/reports/components/report-sections/full-rule-detail.tsx
